refactor: Token父类加上text属性,并将String类成员property删除,下放到各个子类,子类NumToken中property改成double类,并相应地修改LexerTest中的输出方法.

+++ src/com/whc/lexer/Lexer.java
@@ -68,7 +68,7 @@ public class Lexer {
                 c = getChar();
                 if(!isDigit(c)){
                     ungetChar(c);
-                    return new NumToken(Integer.valueOf(word.toString()),lineIndex);
+                    return new NumToken(Double.valueOf(word.toString()),lineIndex);
                 }
             }while (isDigit(c));
         }else if(isLetter(c)){
